+++ esp32_wireless_control/firmware/msx_motor_driver.cpp
@@ -1,15 +1,15 @@
 #include <Arduino.h>
+
 #include "msx_motor_driver.h"
 #include "uart.h"
 
-MSxMotorDriver::MSxMotorDriver(uint8_t ms1Pin, uint8_t ms2Pin):
-  _ms1Pin(ms1Pin), _ms2Pin(ms2Pin)
+MSxMotorDriver::MSxMotorDriver(uint8_t ms1Pin, uint8_t ms2Pin) : _ms1Pin(ms1Pin), _ms2Pin(ms2Pin)
 {
-  pinMode(_ms1Pin, OUTPUT);
-  pinMode(_ms2Pin, OUTPUT);
+    pinMode(_ms1Pin, OUTPUT);
+    pinMode(_ms2Pin, OUTPUT);
 }
 
-void MSxMotorDriver::setMicrosteps(uint8_t microsteps)
+void MSxMotorDriver::setMicrosteps(uint16_t microsteps)
 {
     switch (microsteps)
     {
