+++ esp32_wireless_control/firmware/msx_motor_driver.h
@@ -3,16 +3,16 @@
 
 #include "motor_driver.h"
 
-class MSxMotorDriver: public MotorDriver
+class MSxMotorDriver : public MotorDriver
 {
-public:
-  MSxMotorDriver(uint8_t ms1Pin, uint8_t ms2Pin);
-  ~MSxMotorDriver() {};
-  void setMicrosteps(uint8_t microsteps);
+  public:
+    MSxMotorDriver(uint8_t ms1Pin, uint8_t ms2Pin);
+    ~MSxMotorDriver(){};
+    void setMicrosteps(uint16_t microsteps);
 
-private:
-  uint8_t _ms1Pin;
-  uint8_t _ms2Pin;
+  private:
+    uint8_t _ms1Pin;
+    uint8_t _ms2Pin;
 };
 
 #endif /* _MSX_MOTOR_DRIVER_H_ */
